.20325203252*1049 != .66666666667*319   rounding error nailed

+++ utils/imutils.py
@@ -855,17 +855,17 @@ def show_mask_with_labels(mask_filename,labels,original_image=None):
             if height>400:
                 print('got a big one (hxw {}x{}) resizing'.format(height,width))
                 factor = 400.0/height
-                orig_arr = cv2.resize(orig_arr,(round(width*factor),400))
+                orig_arr = cv2.resize(orig_arr,(int(round(width*factor)),400))
                 print('factor {} newsize {}'.format(factor,orig_arr.shape) )
 
                 colorbar_h,colorbar_w = dest_colorbar.shape[0:2]
                 factor = 400.0/colorbar_h
-                dest_colorbar = cv2.resize(dest_colorbar,(round(colorbar_w*factor),round(colorbar_h*factor)))
+                dest_colorbar = cv2.resize(dest_colorbar,(int(round(colorbar_w*factor)),int(round(colorbar_h*factor))))
                 print('cbarfactor {} newsize {}'.format(factor,dest_colorbar.shape) )
 
                 dest_h,dest_w = dest.shape[0:2]
                 factor = 400.0/dest_h
-                dest = cv2.resize(dest,(round(dest_w*factor),round(dest_h*factor)))
+                dest = cv2.resize(dest,(int(round(dest_w*factor)),int(round(dest_h*factor))))
                 print('maskfactor {} newsize {}'.format(factor,dest.shape) )
 
             cv2.imshow('original',orig_arr)
